code review feedback
- move the complex sempaphore protecting logic to the app
